pyproject.toml
[tool.poetry]
name = "manjusaka-utils"
version = "0.1.0"
description = ""
authors = ["Manjusaka <me@manjusaka.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
pydantic = "^2.3.0"
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
